Refactor python library structure and docs.
