+++ docs/index.rst
@@ -20,4 +20,4 @@ include
 
     miairr/overview
     formats/overview
-    formats/api
+    lang/python/overview
