+++ docs/lang/python/news.rst
@@ -1,0 +1,3 @@
+.. _PythonNews:
+
+.. include:: ../../../lang/python/NEWS.rst
