+++ docs/lang/python/overview.rst
@@ -1,0 +1,14 @@
+.. _PythonOverview:
+
+Python Library
+=======================
+.. toctree::
+    :maxdepth: 3
+    :hidden:
+
+    api
+    news
+    tools
+
+.. include:: ../../../lang/python/README.rst
+
