+++ docs/lang/python/tools.rst
@@ -1,0 +1,7 @@
+.. _PythonTools:
+
+Commandline Tools
+-------------------------------
+
+.. autoprogram:: airr.tools:define_args()
+   :prog: airr-tools
