+++ docs/requirements.txt
@@ -1,5 +1,5 @@
 sphinx>=1.4
-sphinxcontrib-autoprogram>=0.1.2
+sphinxcontrib-autoprogram>=0.1.4
 prov>=1.5.1
 pyyaml>=3.12
 recommonmark>=0.4.0
