+++ lang/python/NEWS.rst
@@ -1,0 +1,7 @@
+Release Notes
+===============================================================================
+
+Version 1.1.0:  April 20, 2018
+-------------------------------------------------------------------------------
+
+Initial release.
