+++ lang/python/README.rst
@@ -1,14 +1,5 @@
-AIRR Python API
-=====================
-
 Python reference library for AIRR schema for Ig/TCR rearrangements.
 
-Documentation is available on
-`readthedocs <https://airr-formats.readthedocs.org>`__.
-
-Installation
---------------------
-
-In the usual manner::
+Install the usual manner::
 
     python setup.py install --user
